requirements.txt
bleach==3.2.1
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
colorama==0.4.4
docutils==0.16
google-api-core==1.23.0
google-api-python-client==1.12.8
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
importlib-metadata==3.1.1
keyring==21.5.0
oauthlib==3.1.0
packaging==20.7
pkginfo==1.6.1
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.7.3
pyparsing==2.4.7
pytz==2020.4
PyYAML==5.3.1
readme-renderer==28.0
requests==2.25.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
rsa==4.6
six==1.15.0
tqdm==4.54.1
twine==3.2.0
uritemplate==3.0.1
urllib3==1.26.2
webencodings==0.5.1
zipp==3.4.0
